feat: replace NavbarAuth with NavbarGuest and update Sidebar layout
